Some QOL fixes (#659)
* PReLU forward works-ish

* Doctest works now

* Better-ish implementation

PReLU forward and backward now working properly

* Now has actual layers to use

* fmt, clippy, tests

* Cleaning

* Remove one unneeded generic

* Cuda maybe working (idk)

Actually runs now

* LeakyReLU now should work

Changed some defaults to match Pytorch

* Fix test

* Conforms to standards

* Some requested changes

* Formatting

* Clippy

* another fmt

* No prelu kernel

* Fmt

* Explicit zero

* fmt

* Actually exports needed things

* Fixes

* fmt

* Fix nighly reature

* Separate into own file

* Better error message

* Actual better merge

* fmt

---------
